+++ golem/ethereum/exceptions.py
