+++ tests/golem/ethereum/test_paymentprocessor.py
